+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 This changelog is used to track all major changes to django-apscheduler.
 
-## v0.5.3 (UNRELEASED)
+## v0.6.0 (UNRELEASED)
 
 **Fixes**
 
@@ -14,6 +14,11 @@ This changelog is used to track all major changes to django-apscheduler.
 - The Django admin page will now show a list of all the manually triggered jobs that could not be completed
   before `settings.APSCHEDULER_RUN_NOW_TIMEOUT` seconds elapsed.
 - Make more of the string output on the admin page Django-translatable.
+- Introduce a `retry_on_db_operational_error` utility decorator for retrying database-related operations when
+  a `django.db.OperationalError` is encountered (
+  Resolves [#145](https://github.com/jcass77/django-apscheduler/issues/145)).
+- Introduce a `ensure_old_connections_are_closed` utility decorator that can be used in jobs that require a fresh
+  connection to the database to always be available.
 
 ## v0.5.2 (2021-01-28)